refactor: pull config into a module
